+++ lib/MForm/Utils/MFormPageHelper.php
@@ -5,6 +5,15 @@
  * @license MIT
  */
 
+namespace MForm\Utils;
+
+
+use rex_exception;
+use rex_fragment;
+use rex_i18n;
+use rex_path;
+use rex_string;
+
 class MFormPageHelper
 {
     /**